requirements.txt
asyncpg
bcrypt
databases
email_validator
fastapi
httpx
orjson
python-dotenv
redis
structlog
uvicorn
